+++ server/src/main/java/com/hexacore/tayo/car/CarService.java
@@ -1,5 +1,6 @@
 package com.hexacore.tayo.car;
 
+import com.hexacore.tayo.car.carRepository.CarRepository;
 import com.hexacore.tayo.car.dto.CreateCarRequestDto;
 import com.hexacore.tayo.car.dto.GetCarResponseDto;
 import com.hexacore.tayo.car.dto.SearchCarsDto;
@@ -15,7 +16,7 @@ import com.hexacore.tayo.category.SubcategoryRepository;
 import com.hexacore.tayo.category.model.Subcategory;
 import com.hexacore.tayo.common.errors.ErrorCode;
 import com.hexacore.tayo.common.errors.GeneralException;
-import com.hexacore.tayo.reservation.ReservationRepository;
+import com.hexacore.tayo.reservation.repository.ReservationRepository;
 import com.hexacore.tayo.reservation.model.Reservation;
 import com.hexacore.tayo.reservation.model.ReservationStatus;
 import com.hexacore.tayo.user.model.User;
