+++ server/src/main/java/com/hexacore/tayo/car/carRepository/CarRepository.java
@@ -1,4 +1,4 @@
-package com.hexacore.tayo.car;
+package com.hexacore.tayo.car.carRepository;
 
 import com.hexacore.tayo.car.model.Car;
 
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarRepository extends JpaRepository<Car, Long>, CarCustomRepository {
+public interface CarRepository extends JpaRepository<Car, Long>, CustomCarRepository {
 
     Optional<Car> findByIdAndIsDeletedFalse(Long id);
