+++ server/src/main/java/com/hexacore/tayo/car/carRepository/CustomCarRepository.java
@@ -1,11 +1,11 @@
-package com.hexacore.tayo.car;
+package com.hexacore.tayo.car.carRepository;
 
 import com.hexacore.tayo.car.dto.SearchCarsDto;
 import com.hexacore.tayo.car.dto.SearchCarsResultDto;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
-public interface CarCustomRepository {
+public interface CustomCarRepository {
 
     Slice<SearchCarsResultDto> search(SearchCarsDto searchCarsDto, Pageable pageable);
